requirements.txt
django==2.2
django-cors-headers==3.0.2
django-filter==2.0.0
django-fixture-magic==0.1.3
django-rest-swagger==2.2.0
djangorestframework==3.8.2
future==0.17.1
psycopg2>=2.7,<3.0
ipython==7.2.0
pyyaml==3.13